+++ packages/common/web/src/Utils/contexts.tsx
@@ -1,11 +1,13 @@
 import React from 'react';
-import { TranslationFunction, AvailableExperiments } from '../types';
+import { TranslationFunction, AvailableExperiments, TextDirection } from '../types';
 
 export const GlobalContext = React.createContext<{
   experiments: AvailableExperiments;
   isMobile: boolean;
   t?: TranslationFunction;
+  languageDir?: TextDirection;
 }>({
   isMobile: false,
   experiments: {},
+  languageDir: 'ltr',
 });